### 2021-03-14 #3108 @whedon
[PRE REVIEW]: ngsxfem: Add-on to NGSolve for geometrically unfitted finite element discretizations
**Submitting author:** @schruste (<a href="http://orcid.org/0000-0003-0170-8468">Christoph Lehrenfeld</a>)
**Repository:** <a href="https://github.com/ngsxfem/ngsxfem" target ="_blank">https://github.com/ngsxfem/ngsxfem</a>
**Version:** v2.0
