### 2021-03-14 a9ddc09
add missing dois (thanks @whedon)
