+++ doc/lit-ngsxfem.bib
@@ -86,6 +86,7 @@
   eprinttype  = {arXiv},
   keywords    = {math.NA, 65N30, 65N12, 65N15},
   timestamp   = {2021-01-05},
+  doi         = {10.1093/imanum/drz062},
 }
 
 @Online{JR19a,
@@ -247,6 +248,7 @@
   eprinttype  = {arXiv},
   keywords    = {math.NA},
   timestamp   = {2021-01-05},
+  doi         = {10.7712/100016.1820.4573},
 }
 
 @Article{GLR18,
