### 2021-03-17 8beb543
GCC sparsity pattern debugging
