+++ spacetime/py_demos/demos_spacetime_spacetimeGCC_debug3.py
@@ -1,0 +1,229 @@
+"""
+unfitted Heat equation with Neumann b.c. solved with an unfitted isoparametric
+space-time discretisation.
+"""
+
+# ------------------------------ LOAD LIBRARIES -------------------------------
+from ngsolve import *
+from netgen.geom2d import SplineGeometry
+from xfem import *
+from math import pi
+from xfem.lset_spacetime import *
+ngsglobals.msg_level = 1
+
+# -------------------------------- PARAMETERS ---------------------------------
+
+# DISCRETIZATION PARAMETERS:
+# parameter for refinement study:
+i = 2
+n_steps = 2**i
+space_refs = i+1
+
+# polynomial order in time
+k_t = 3
+# polynomial order in space
+k_s = 0 #k_t
+# polynomial order in time for level set approximation
+lset_order_time = k_t
+# integration order in time
+time_order = 2 * k_t
+# time stepping parameters
+tstart = 0
+tend = 0.5
+delta_t = (tend - tstart) / n_steps
+maxh = 10.5
+# ghost penalty parameter
+gamma = 0.05
+# map from reference time to physical time
+told = Parameter(tstart)
+t = told + delta_t * tref
+
+# PROBLEM SETUP:
+
+# outer domain:
+rect = SplineGeometry()
+rect.AddRectangle([-1, -1], [1, 1])
+
+# level set geometry
+# radius of disk (the geometry)
+R = 0.5
+# position shift of the geometry in time
+rho = (1 / (pi)) * sin(2 * pi * t)
+# convection velocity:
+w = CoefficientFunction((0, rho.Diff(t)))
+max_velocity = 2
+# level set
+r = sqrt(x**2 + (y - rho)**2)
+levelset = r - R
+
+# diffusion coeff
+alpha = 1
+# solution
+#u_exact = cos(pi * r / R) * (1-cos(pi * t))
+u_exact = cos(pi*x)*cos(pi*y)* (1-cos(pi * t))
+#u_exact = (1-cos(pi * t))
+# r.h.s.
+coeff_f = (u_exact.Diff(t)
+           - alpha * (u_exact.Diff(x).Diff(x) + u_exact.Diff(y).Diff(y))
+           + w[0] * u_exact.Diff(x) + w[1] * u_exact.Diff(y)).Compile()
+#coeff_f = 2*t
+
+# ----------------------------------- MAIN ------------------------------------
+
+ngmesh = rect.GenerateMesh(maxh=maxh, quad_dominated=True)
+#for j in range(space_refs):
+    #ngmesh.Refine()
+mesh = Mesh(ngmesh)
+
+# spatial FESpace for solution
+fes = L2(mesh, order=k_s, dgjumps=True)
+# time finite elements (nodal!)
+tfe_i = GCC3FE(skip_first_nodes=True)  # interior shapes
+tfe_e = GCC3FE(only_first_nodes=True)  # exterior shapes (init. val.)
+tfe_t = ScalarTimeFE(0)                    # test shapes
+# space-time finite element space
+
+st_fes_i = tfe_i * fes
+st_fes_e = tfe_e * fes
+st_fes_t = tfe_t * fes
+
+fes_t = st_fes_t * fes 
+
+u_i = st_fes_i.TrialFunction()
+v_t, w_t  = fes_t.TestFunction()
+
+# Space time version of Levelset Mesh Adaptation object. Also offers integrator
+# helper functions that involve the correct mesh deformation
+lsetadap = LevelSetMeshAdaptation_Spacetime(mesh, order_space=k_s,
[…]
