+++ demos/spacetime/spacetimeDG_fitted_paperstuff.py
@@ -1,117 +1,0 @@
-"""
-Heat equation (fitted) with Dirichlet b.c. solved with a DG-in-time
-space-time finite element discretisation
-
-Domain:
--------
-
-The background domain is [-0.6,0.6]x[-1,1]x[0,0.5] (2D+time interval)
-while the physical domain is a circle that is traveling up and down
-over time.
-
-PDE problem:
-------------
-  u_t - (u_xx + u_yy) = f in  Omega=[0,1]^2
-                    u = 0 on dOmega
-The r.h.s. term f is chosen according to a manufactured solution.
-
-Discretisation:
----------------
-* Discontinuous-in-time Galerkin space-time finite elements
-
-Implementational aspects:
--------------------------------------------------------
-* A (sparse) direct solver is applied to solve the arising linear systems.
-
-References:
------------
-All concepts that are used here are explained in the jupyter-tutorial
-`spacetime.ipynb` (for the more involved case of moving domains).
-"""
-
-# ------------------------------ LOAD LIBRARIES -------------------------------
-from netgen.geom2d import SplineGeometry
-from ngsolve import *
-from xfem import *
-from math import pi
-ngsglobals.msg_level = 1
-
-# -------------------------------- PARAMETERS ---------------------------------
-order = 1
-k_t = 1
-tend = 1.0
-delta_t = 1 / 32
-
-
-# ----------------------------------- MAIN ------------------------------------
-
-geo = SplineGeometry()
-geo.AddCircle(c=(0,0),
-              r=0.5,
-              bc="circle",
-              leftdomain=2,
-              rightdomain=0)
-
-mesh = Mesh(geo.GenerateMesh(maxh=0.2, quad_dominated=False))
-mesh.Curve(3)
-
-outfile = open("tikz_mesh_fitted.tex", "w")
-name = "trig"
-
-for t in mesh.ngmesh.Elements2D():
-    p1 = mesh.ngmesh.Points()[t.points[0]]
-    p2 = mesh.ngmesh.Points()[t.points[1]]
-    p3 = mesh.ngmesh.Points()[t.points[2]]
-    
-    print(mesh.deformation(p1))
-                
-    outfile.write("\\draw["+name+"] ( "+str(p1[0])+" , "+str(p1[1])+" ) -- ( "+str(p2[0])+" , "+str(p2[1])+" ) -- ( "+str(p3[0])+" , "+str(p3[1])+" ) -- cycle;\n")
-outfile.close()
-
-V = H1(mesh, order=order, dirichlet=".*")
-tfe = ScalarTimeFE(k_t)
-st_fes = tfe * V
-
-# Fitted heat equation example
-tnew = 0
-told = Parameter(0)
-t = told + delta_t * tref
-
-u_exact = sin(pi * t) * sin(pi * x)**2 * sin(pi * y)**2
-
-coeff_f = u_exact.Diff(t) - (u_exact.Diff(x).Diff(x) + u_exact.Diff(y).Diff(y))
-coeff_f = coeff_f.Compile()
-
-gfu = GridFunction(st_fes)
-u_last = CreateTimeRestrictedGF(gfu, 1)
-u, v = st_fes.TnT()
-
-dxt = delta_t * dxtref(mesh, time_order=2)
-dxold = dmesh(mesh, tref=0)
-dxnew = dmesh(mesh, tref=1)
-
-
-def dt(u): return 1.0 / delta_t * dtref(u)
-
-
-a = BilinearForm(st_fes, symmetric=False)
-a += grad(u) * grad(v) * dxt
-a += u * v * dxold
[…]
