### 2021-03-18 d848fb3
Making GCC static for debugging
