### 2021-03-18 518ba83
fixed GCC example
