+++ utils/restrictedblf.cpp
@@ -17,122 +17,158 @@ namespace ngcomp
     ;
   }
   
+    //associate elements / specialelements / facets (DG) to dofs for precomputed sparsity pattern
+  Table<int> MeshEntityToDofTable(shared_ptr<FESpace> fes, 
+                                  shared_ptr<BitArray> active_elems = nullptr,
+                                  shared_ptr<BitArray> active_facets = nullptr,
+                                  bool eliminate_internal = false, 
+                                  bool eliminate_hidden = false, 
+                                  const Array<unique_ptr<SpecialElement>> * specels = nullptr)
+  {
+    if (eliminate_internal)
+      if (fes->UsesDGCoupling())
+        cout << IM(4) << "static condensation should work here, but the user should make sure that all local dofs are not involved in dg-couplings." << endl;
+    if (eliminate_hidden)
+      if (fes->UsesDGCoupling())
+        cout << IM(4) << "eliminating hidden should work here, but the user should make sure that the hidden dofs are not involved in dg-couplings." << endl;
+
+    shared_ptr<MeshAccess> ma = fes->GetMeshAccess();
+    
+    size_t ndof = fes->GetNDof();
+    size_t nf = ma->GetNFacets();
+    size_t neV = ma->GetNE(VOL);
+    size_t neB = ma->GetNE(BND);
+    size_t neBB = ma->GetNE(BBND);
+    size_t nspe = 0;
+    if (specels)
+      nspe = specels->Size();
+    
+    int maxind = neV + neB + neBB + nspe;
+    if (fes->UsesDGCoupling()) 
+      maxind += nf;
+
+    TableCreator<int> creator(maxind);
+
+    for ( ; !creator.Done(); creator++)
+    {
+      for(VorB vb : {VOL, BND, BBND})
+      {
+        size_t shift = (vb==VOL) ? 0 : ((vb==BND) ? neV : neV+neB);
+        bool condensation_allowed = (vb == VOL) || ((neV==0) && (vb == BND));
+	      ParallelForRange (ma->GetNE(vb), [&](IntRange r)
+	      {
+          Array<DofId> dnums;
+          for (auto i : r)
+          {
+            if (vb == VOL)
+              if (active_elems && (! active_elems->Test(i)))
+                continue;
+            auto eid = ElementId(vb,i);
+            if (!fes->DefinedOn (vb, ma->GetElIndex(eid))) continue;
+            if (condensation_allowed && eliminate_internal)
+              fes->GetDofNrs (eid, dnums, EXTERNAL_DOF);
+            else if (condensation_allowed && eliminate_hidden)
+              fes->GetDofNrs (eid, dnums, VISIBLE_DOF);
+            else
+              fes->GetDofNrs (eid, dnums);
+            for (DofId d : dnums)
+              if (IsRegularDof(d)) creator.Add (shift+i, d);
+          }
+        });
+      }
+      if (specels) 
+      {
+        Array<DofId> dnums;
+        for (int i = 0; i < specels->Size(); i++)
+        {
+          (*specels)[i]->GetDofNrs (dnums);
+          QuickSort(dnums);
+          int last = -1;
+          for (int d : dnums)
+          {
+            if (d!=last && IsRegularDof(d))
+              creator.Add (neV+neB+neBB+i, d);
+            last = d;
+          }
+        }
+      }
+
+      if (fes->UsesDGCoupling())
+      {
+	      ParallelForRange (nf, [&](IntRange r)
+	      {
+          //add dofs of neighbour elements as well
+          //by associating those to facets
+          Array<DofId> dnums;
+          Array<DofId> dnums_dg;
+          Array<int> elnums;
+          Array<int> elnums_per;
+          Array<int> nbelems;
[…]
