+++ .gitlab-ci.yml
@@ -5,7 +5,7 @@ stages:
         
 build:
   stage: build
-  image: schruste/ngsolve:latest
+  image: ngsxfem/ngsolve:latest
   script:
     - git submodule update --init --recursive
     - mkdir build
