### 2021-04-21 4a0e3e1
support 1D (at least for fitted space-time)
