### 2021-04-21 5259bd5
adjust DiffShape to newest ngsolve (Eulerian flag)
