+++ utils/restrictedfespace.hpp
@@ -159,7 +159,8 @@ namespace ngfem
               BareSliceVector<Complex> x) const override;
 
     shared_ptr<CoefficientFunction> DiffShape (shared_ptr<CoefficientFunction> proxy,
-                                               shared_ptr<CoefficientFunction> dir) const override;
+                                               shared_ptr<CoefficientFunction> dir,
+                                               bool Eulerian) const override;
   };
 
 
