### 2021-05-11 866f47d
move literature to separate md file
