+++ doc/literature.md
@@ -1,0 +1,69 @@
+The library `ngsxfem` has been used in the following scientific literature:
+
+#### Scalar fictitious domain problems
+* A higher order isoparametric fictitious domain method for level set domains. In S. Bordas, E. Burman, M. Larson, & M. Olshanskii (Eds.), Lecture notes in computational science and engineering (Vol. 121, pp. 65–92). Springer, Cham. doi: [10.1007/978-3-319-71431-8_3](https://doi.org/10.1007/978-3-319-71431-8_3).
+
+#### Scalar and Stokes interface problems, such two-phase flow problems
+
+* C. Lehrenfeld and A. Reusken. “Optimal preconditioners for Nitsche-XFEM discretizations of interface problems”. In: Numer. Math. 135.2 (26th Mar. 2016), pp. 313–332. doi: [10.1007/s00211-016-0801-6](https://doi.org/10.1007/s00211-016-0801-6).
+
+* C. Lehrenfeld. “High order unfitted finite element methods on level set domains using isoparametric mappings”. In: Comput. Methods Appl. Mech. Engrg. 300 (1st Mar. 2016), pp. 716–733. doi: [10.1016/j.cma.2015.12.005](https://doi.org/10.1016/j.cma.2015.12.005).
+
+* C. Lehrenfeld. “Removing the stabilization parameter in fitted and unfitted symmetric Nitsche formulations”. In: Proceedings of the VII European Congress on Computational Methods in Applied Sciences and Engineering (ECCOMAS Congress 2016). Institute of Structural Analysis and Antiseismic Research School of Civil Engineering National Technical University of Athens (NTUA) Greece, 2nd Mar. 2016. doi: [10.7712/100016.1820.4573](https://doi.org/10.7712/100016.1820.4573).
+
+* C. Lehrenfeld and A. Reusken. “L2-error analysis of an isoparametric unfitted finite element method for elliptic interface problems”. In: J. Numer. Math. 27.2 (17th Dec. 2019), pp. 85–99. doi: [10.1515/jnma-2017-0109](https://doi.org/10.1515/jnma-2017-0109).
+
+* P. Lederer, C.-M. Pfeiler, C. Wintersteiger and C. Lehrenfeld. “Higher order unfitted FEM for Stokes interface problems”. In: PAMM. Proc. Appl. Math. Mech. 16.1 (25th Oct. 2016), pp. 7–10. doi: [10.1002/pamm.201610003](https://doi.org/10.1002/pamm.201610003).
+
+* M. Olshanskii, A.Quaini and Q. Sun. An unfitted finite element method for two-phase Stokes problems with slip between phases. 24th Jan. 2021. arXiv: [2101.09627](https://arxiv.org/abs/2101.09627) [math.NA].
+
+* T. Ludescher. “Multilevel preconditioning of stabilized unfitted finite element discretizations”. PhD thesis. RWTH Aachen University, 27th May 2020. doi: [10.18154/RWTH-2020-07305](https://doi.org/10.18154/RWTH-2020-07305.)
+
+#### PDEs on moving domains: Space-time methods
+
+* J. Preuß. “Higher order unfitted isoparametric space-time FEM on moving domains”. MA thesis. Georg-August Universität Göttingen, 31st Jan. 2018. url: http://cpde.math.uni-goettingen.de/data/Pre18_Ma.pdf.
+
[…]
