+++ doc/lit-ngsxfem.bib
@@ -9,9 +9,8 @@
   eprinttype  = {arXiv},
 }
 
-
 @Online{OQS21,
-  author      = {M. Olshanskii and A.Quaini and Q. Sun},
+  author      = {M. Olshanskii and A. Quaini and Q. Sun},
   date        = {2021-01-24},
   title       = {An unfitted finite element method for two-phase {Stokes} problems with slip between phases},
   eprint      = {2101.09627},
@@ -41,15 +40,17 @@
   timestamp = {2021-01-05},
 }
 
-@Online{vWRL20,
-  author      = {H. von Wahl and T. Richter and C. Lehrenfeld},
-  date        = {2020-02-06},
-  title       = {An unfitted {Eulerian} finite element method for the time-dependent {Stokes} problem on moving domains},
-  eprint      = {2002.02352},
-  eprintclass = {math.NA},
-  eprinttype  = {arXiv},
-  keywords    = {math.NA, cs.NA},
-  timestamp   = {2021-01-05},
+@Article{vWRL20,
+  author       = {H. von Wahl and T. Richter and C. Lehrenfeld},
+  date         = {2021},
+  journaltitle = {IMA J. Numer. Anal.},
+  title        = {An unfitted {Eulerian} finite element method for the time-dependent {Stokes} problem on moving domains},
+  doi          = {10.1093/imanum/drab044},
+  eprint       = {2002.02352},
+  eprintclass  = {math.NA},
+  eprinttype   = {arXiv},
+  keywords     = {math.NA, cs.NA},
+  timestamp    = {2021-01-05},
 }
 
 @Dataset{vWRL20a,
@@ -77,16 +78,23 @@
   timestamp    = {2021-01-05},
 }
 
-@Online{JR19,
-  author      = {T. Jankuhn and A. Reusken},
-  date        = {2019-04-29},
-  title       = {Trace Finite Element Methods for Surface Vector-{Laplace} Equations},
-  eprint      = {1904.12494},
-  eprintclass = {math.NA},
-  eprinttype  = {arXiv},
-  keywords    = {math.NA, 65N30, 65N12, 65N15},
-  timestamp   = {2021-01-05},
-  doi         = {10.1093/imanum/drz062},
+@Article{JR19,
+  author       = {T. Jankuhn and A. Reusken},
+  date         = {2019-04-29},
+  journaltitle = {{IMA} Journal of Numerical Analysis},
+  title        = {Trace finite element methods for surface vector-Laplace equations},
+  doi          = {10.1093/imanum/drz062},
+  eprint       = {1904.12494},
+  eprintclass  = {math.NA},
+  eprinttype   = {arXiv},
+  number       = {1},
+  pages        = {48--83},
+  volume       = {41},
+  keywords     = {math.NA, 65N30, 65N12, 65N15},
+  month        = may,
+  publisher    = {Oxford University Press ({OUP})},
+  timestamp    = {2021-01-05},
+  year         = {2020},
 }
 
 @Online{JR19a,
@@ -239,16 +247,19 @@
   timestamp   = {2021-01-05},
 }
 
-@Online{Leh16a,
+@InProceedings{Leh16a,
   author      = {C. Lehrenfeld},
+  booktitle   = {Proceedings of the {VII} European Congress on Computational Methods in Applied Sciences and Engineering ({ECCOMAS} Congress 2016)},
   date        = {2016-03-02},
   title       = {Removing the stabilization parameter in fitted and unfitted symmetric {Nitsche} formulations},
+  doi         = {10.7712/100016.1820.4573},
   eprint      = {1603.00617},
   eprintclass = {math.NA},
   eprinttype  = {arXiv},
[…]
