+++ doc/paper.md
@@ -32,7 +32,7 @@ bibliography: lit-ngsxfem.bib
 # Summary
 `ngsxfem` is an add-on library to [`Netgen/NGSolve`](www.ngsolve.org), a general purpose, high performance finite element library for the numerical solution of partial differential equations. The add-on enables the use of geometrically unfitted finite element technologies known under different labels, e.g. *XFEM*, *CutFEM*, *TraceFEM*, *Finite Cell*, *fictitious domain method* or *Cut-Cell methods*, etc.. Both, `Netgen/NGSolve` and `ngsxfem` are written in C++ with a rich python interface through which it is typically used. 3`ngsxfem` is an academic software. Its primary intention is to facilitate the development and validation of new numerical methods.
 
-# Scope and purpose
+# Statement of need
 Typically, in the finite element method for the discretization of PDEs, the geometry under consideration is parametrized by the computational mesh yielding *geometrically fitted* finite element methods. The generation and adaptation of geometrically fitted computational meshes can be a burden on simulation methods, e.g. if the geometries are complex or especially if they are evolving in time. To be more flexible w.r.t. the geometry, *geometrically unfitted* finite element methods can be considered which break the direct link between the geometry parametrization and the computational mesh. Instead, a separate description of the geometry, e.g. through a *level set function* is used. `ngsxfem` aims at providing the necessary tools to robustly work in a *geometrically unfitted* setting where the geometry is described by one (or multiple) *level set function(s)*. The essential tools extending standard finite element codes for the *geometrically unfitted* setting are:
 
 * formulation of geometrically unfitted geometry through level set function(s)
@@ -47,45 +47,8 @@ First of all `ngsxfem` provides these tools for `Netgen/NGSolve`. For other fini
 * space-time finite elements and quadrature for unfitted space-time finite element discretizations of PDEs on moving domains, cf. @Pre18, @Hei20
 * the so-called direct version of the ghost penalty stabilization as introduced in @Pre18
 
-A detailed overview of the scientific literature which has utilised `ngsxfem` is given in the repository under `docu/literature.md`
+A more detailed overview of the key features provided by `ngsxfem` can be found in the repository under `docu/feature-details.md` Furthermore, a detailed overview of the scientific literature which has utilised `ngsxfem` is given in the repository under `docu/literature.md`
 
-## Feature details
-In the previous section we briefly explained or mentioned some tools of `ngsxfem`. In this section we elaborate on a few non-trivial aspects of them.
-
-### Tools to work on an "active mesh" only
-In unfitted finite element methods some functions and integrals are only defined on a subset of the mesh. Accordingly finite element spaces and integrals have to be defined only on this active part of the mesh. 
-`ngsxfem` offers the tools to mark the corresponding elements and facets and use the marking during assembly and definition of finite element spaces. 
-On cut elements one often also uses locally modified finite elements, e.g. by restriction of finite elements on the background mesh.
-
-![Left: Active mesh. Right: Basis function restricted to a subdomain.](graphics/xfes.png){ height=3cm align=center}
-
-### Numerical integration on unfitted geometries described by one level set function
-Given a level set function $\phi$ which describes the geometry (e.g. $\Omega = \{ \phi < 0 \}$), a piecewise linear (or bilinear on hyperrectangles, cf. @HL19) approximation is made. How to obtain a higher order reconstruction from this basis approximation is discussed [below](#higher-order-representation-of-implicit-level-set-geometries).
-
-On simplices (triangles and tetrahedra) this gives a planar intersection on every element which allows for an explicit decomposition into simple geometries.
[…]
