+++ doc/literature.md
@@ -1,9 +1,11 @@
-The library `ngsxfem` has been used in the following scientific literature:
+---
+title: 'Scientific literature using `ngsxfem`'
+---
 
-#### Scalar fictitious domain problems
+## Scalar fictitious domain problems
 * A higher order isoparametric fictitious domain method for level set domains. In S. Bordas, E. Burman, M. Larson, & M. Olshanskii (Eds.), Lecture notes in computational science and engineering (Vol. 121, pp. 65–92). Springer, Cham. doi: [10.1007/978-3-319-71431-8_3](https://doi.org/10.1007/978-3-319-71431-8_3).
 
-#### Scalar and Stokes interface problems, such two-phase flow problems
+## Scalar and Stokes interface problems, such as two-phase flow problems
 
 * C. Lehrenfeld and A. Reusken. “Optimal preconditioners for Nitsche-XFEM discretizations of interface problems”. In: Numer. Math. 135.2 (26th Mar. 2016), pp. 313–332. doi: [10.1007/s00211-016-0801-6](https://doi.org/10.1007/s00211-016-0801-6).
 
@@ -19,25 +21,25 @@ The library `ngsxfem` has been used in the following scientific literature:
 
 * T. Ludescher. “Multilevel preconditioning of stabilized unfitted finite element discretizations”. PhD thesis. RWTH Aachen University, 27th May 2020. doi: [10.18154/RWTH-2020-07305](https://doi.org/10.18154/RWTH-2020-07305.)
 
-#### PDEs on moving domains: Space-time methods
+## PDEs on moving domains: Space-time methods
 
 * J. Preuß. “Higher order unfitted isoparametric space-time FEM on moving domains”. MA thesis. Georg-August Universität Göttingen, 31st Jan. 2018. url: http://cpde.math.uni-goettingen.de/data/Pre18_Ma.pdf.
 
 * F. Heimann. “On Discontinuous- and Continuous-In-Time Unfitted Space-Time Methods for PDEs on Moving Domains”. MA thesis. Georg-August Universität Göttingen, Oct. 2020. url: http://cpde.math.uni-goettingen.de/data/Hei20_Ma.pdf.
 
-#### PDEs on moving domains: Eulerian time-stepping
+## PDEs on moving domains: Eulerian time-stepping
 * C. Lehrenfeld and M. A. Olshanskii. “An Eulerian finite element method for PDEs in time-dependent domains”. In: ESAIM Math. Model. Numer. Anal. 53.2 (24th Apr. 2019), pp. 585–614. doi: [10.1051/m2an/2018068](https://doi.org/10.1051/m2an/2018068).
 
 * H. von Wahl, T. Richter and C. Lehrenfeld. “An unfitted Eulerian finite element method for the time-dependent Stokes problem on moving domains”. In: IMA J. Numer. Anal. (2021). doi: [10.1093/imanum/drab044](https://doi.org/10.1093/imanum/drab044). arXiv: [2002.02352](https://arxiv.org/abs/2002.02352) [math.NA].
 
-#### Fluid structure interaction
+## Fluid structure interaction
 
 * H. von Wahl, T. Richter, S. Frei and T. Hagemeier. “Falling balls in a viscous fluid with contact: Comparing numerical simulations with experimental data”. In: Phys. Fluids 33.3, 033304 (2nd Mar. 2021). doi: [10.1063/5.0037971](https://doi.org/10.1063/5.0037971).
 
[…]
