### 2021-05-12 e8b2760
link to artifacts in README and fix a literature entry
