### 2021-05-12 #3237 @schruste
Hi @mikaem ,

Thank you for the feedback. It seems we actually overlooked the guideline on the length at first, but repaired this now. We moved the more detailed parts to separate markdown files in `doc` (which shouldn't be directly relevant for the review anymore).

As for the docker question. We added a sentence in the `INSTALLATION.md` to emphasize that it is really the browser of the host system from which you spawn the docker image (not a browser in the docker!) that can be used (thus the port forwarding in the spawning command). The docker image itself does not need to have a browser.

The updated changes are in the `master` branch (which is not the default branch) for now.

Best,
Christoph
