### 2021-05-18 cb340df
missing doi and url in literature list
