+++ doc/lit-ngsxfem.bib
@@ -370,6 +370,7 @@
   journaltitle = {Kybernetika},
   title        = {{Generic implementation of finite element methods in the Distributed and Unified Numerics Environment (DUNE)}},
   pages        = {294--315},
+  url          = {http://www.kybernetika.cz/content/2010/2/294/paper.pdf},
   volume       = {46},
 }
 
@@ -378,6 +379,7 @@
   date         = {2018},
   journaltitle = {Archives of Computational Methods in Engineering},
   title        = {FEMPAR: An object-oriented parallel finite element framework},
+  doi          = {10.1007/s11831-017-9244-1},
   number       = {2},
   pages        = {195--271},
   volume       = {25},
@@ -391,7 +393,6 @@
   eprint      = {2105.01163},
   eprintclass = {math.NA},
   eprinttype  = {arXiv},
-  file        = {:http\://arxiv.org/pdf/2105.01163v1:PDF},
   keywords    = {math.NA, cs.NA},
   timestamp   = {2021-05-11},
 }
