### 2021-05-20 #3237 @mscroggs
Hi @schruste,

This is a nice piece of software, and a good write up. I successfully ran some demos in the docker image, and also successfully installed xfem using pip (after successfully installing NGSolve using apt).

I left two boxes unchecked as I have a couple of minor changes to suggest:

- (Installation) INSTALLATION.md points to docker image ngsxfem/ngsxfem, but the readme points to docker image schruste/ngsxfem. I think the first link is correct and the readme should be updated?
- (References) I think the following DOIs are valid for the two references currently without DOIs or links:
  - Badia, Martin, Principe: https://dx.doi.org/10.1007/s11831-017-9244-1
  - Bastian, Heimann, Marnach: https://dx.doi.org/10338.dmlcz/140745
