+++ doc/lit-ngsxfem.bib
@@ -369,8 +369,8 @@
   date         = {2010},
   journaltitle = {Kybernetika},
   title        = {{Generic implementation of finite element methods in the Distributed and Unified Numerics Environment (DUNE)}},
+  doi          = {10338.dmlcz/140745},
   pages        = {294--315},
-  url          = {http://www.kybernetika.cz/content/2010/2/294/paper.pdf},
   volume       = {46},
 }
 
