### 2021-05-20 #3237 @schruste
Hi @mscroggs,

Thank you for the nice feedback and for the reports. We added the DOIs and fixed the `README.md`!
