### 2021-05-21 #3237 @mikaem
I had a final reading of the article and I saw two more minor issues. There are some periods missing in the last paragraph before acknowledgements, and I seem to be mentioned twice in the left column under reviewers for some reason?
