### 2021-05-21 #3237 @schruste
Hi @mikaem ,

Thanks for the hint with the period. I don't know why you are mentioned twice, but as far as I see this is not related to our content but because you are listed twice in this issue (`Reviewer: @mscroggs, @mikaem, @mikaem`).
