### 2021-05-25 fb61da6
trying to templatize RestrictedBilinearForm
