+++ utils/python_utils.cpp
@@ -180,8 +180,8 @@ BilinearForm restricted on a set of elements and facets.
           if (py::extract<PyBA> (afac_restriction).check())
             fac_restriction = py::extract<PyBA>(afac_restriction)();
 
-          //if (fes1->IsComplex() || fes2->IsComplex())
-          //  throw Exception("RestrictedBilinearForm not implemented for complex fespace");
+          if (fes1->IsComplex() || fes2->IsComplex())
+            throw Exception("RestrictedBilinearForm not implemented for complex fespace");
 
           auto biform = make_shared<RestrictedBilinearForm<double,double>> (fes1, fes2, aname, el_restriction, fac_restriction, flags);
           biform -> SetCheckUnused (check_unused);
@@ -202,6 +202,126 @@ BilinearForm restricted on a set of elements and facets.
                   &RestrictedBilinearForm<double,double>::SetFacetRestriction, "facet restriction")
 
 ;
+
+  auto rblfC_string = docu_string(R"raw_string(
+A restricted bilinear form is a complex bilinear form with a reduced MatrixGraph
+compared to the usual BilinearForm. BitArray(s) define on which elements/facets entries will be
+created.
+
+Use cases:
+
+ * ghost penalty type stabilization:
+    Facet-stabilization that are introduced only act on a few facets in the mesh. By providing the
+    information on the corresponding facets, these additional couplings will only be introduced
+    where necessary.
+
+ * fictitious domain methods:
+    When PDE problems are only solved on a part of a domain while a finite element space is used
+    that is still defined on the whole domain, a BitArray can be used to mark the 'active' part of
+    the mesh.
+
+Parameters
+
+space (trialspace) : ngsolve.FESpace
+  finite element space on which the bilinear form is defined 
+  (trial space and (if no test space is defined) test space).
+
+testspace : ngsolve.FESpace
+  finite element space on which the bilinear form is defined
+  (test space).
+
+name : string
+  name of the bilinear form
+
+element_restriction : ngsolve.BitArray
+  BitArray defining the 'active mesh' element-wise
+
+facet_restriction : ngsolve.BitArray
+  BitArray defining the 'active facets'. This is only relevant if FESpace has DG-terms (dgjumps=True)
+
+check_unused : boolean
+  Check if some degrees of freedoms are not considered during assembly
+
+flags : ngsolve.Flags
+  additional bilinear form flags
+)raw_string");
+
+  py::class_<RestrictedBilinearForm<Complex,Complex>, shared_ptr<RestrictedBilinearForm<Complex,Complex>>, BilinearForm> rblfC(m, "RestrictedBilinearFormC", docu_string(R"raw_string(
+BilinearForm restricted on a set of elements and facets.
+)raw_string") , py::dynamic_attr());
+    
+
+  rblfC.def(py::init([](shared_ptr<FESpace> fes,
+           const string & aname,
+           py::object ael_restriction,
+           py::object afac_restriction,
+           bool check_unused,
+           py::kwargs kwargs)
+        {
+          auto flags = CreateFlagsFromKwArgs(kwargs);
+
+          shared_ptr<BitArray> el_restriction = nullptr;
+          shared_ptr<BitArray> fac_restriction = nullptr;
+          if (py::extract<PyBA> (ael_restriction).check())
+            el_restriction = py::extract<PyBA>(ael_restriction)();
+
+          if (py::extract<PyBA> (afac_restriction).check())
+            fac_restriction = py::extract<PyBA>(afac_restriction)();
+
+          if (!(fes->IsComplex()))
+            throw Exception("For real fespace use RestrictedBilinearForm");
+
[…]
