### 2021-05-27 7356071
combine real/complex restricted bfi
