### 2021-05-27 4fb82ed
slim down python bindings of RestrictedBilinearForm
