+++ utils/python_utils.cpp
@@ -12,6 +12,125 @@
 #include "../utils/restrictedfespace.hpp"
 
 using namespace ngcomp;
+typedef shared_ptr<BitArray> PyBA;
+
+
+ auto rblf_string_T = docu_string(R"raw_string(
+A restricted bilinear form is a bilinear form with a reduced MatrixGraph
+compared to the usual BilinearForm. BitArray(s) define on which elements/facets entries will be
+created.
+
+Use cases:
+
+ * ghost penalty type stabilization:
+    Facet-stabilization that are introduced only act on a few facets in the mesh. By providing the
+    information on the corresponding facets, these additional couplings will only be introduced
+    where necessary.
+
+ * fictitious domain methods:
+    When PDE problems are only solved on a part of a domain while a finite element space is used
+    that is still defined on the whole domain, a BitArray can be used to mark the 'active' part of
+    the mesh.
+
+Parameters
+
+space (trialspace) : ngsolve.FESpace
+  finite element space on which the bilinear form is defined 
+  (trial space and (if no test space is defined) test space).
+
+testspace : ngsolve.FESpace
+  finite element space on which the bilinear form is defined
+  (test space).
+
+name : string
+  name of the bilinear form
+
+element_restriction : ngsolve.BitArray
+  BitArray defining the 'active mesh' element-wise
+
+facet_restriction : ngsolve.BitArray
+  BitArray defining the 'active facets'. This is only relevant if FESpace has DG-terms (dgjumps=True)
+
+check_unused : boolean
+  Check if some degrees of freedoms are not considered during assembly
+
+flags : ngsolve.Flags
+  additional bilinear form flags
+)raw_string");
+
+template <class TM,class TV>
+void declare_RestrictedBilinearForm(py::module &m, std::string const &typestr) {
+using Rbfi_TT = RestrictedBilinearForm<TM,TV>;
+std::string pyclass_name_T = std::string("RestrictedBilinearForm") + typestr;
+
+py::class_<Rbfi_TT, shared_ptr<Rbfi_TT>, BilinearForm> rblf_T(m, pyclass_name_T.c_str(),docu_string(R"raw_string(BilinearForm restricted on a set of elements and facets.
+)raw_string") , py::dynamic_attr());
+
+rblf_T.def(py::init([](shared_ptr<FESpace> fes,
+      const string & aname,
+      py::object ael_restriction,
+      py::object afac_restriction,
+      bool check_unused,
+      py::kwargs kwargs)
+  {
+    auto flags = CreateFlagsFromKwArgs(kwargs);
+
+    shared_ptr<BitArray> el_restriction = nullptr;
+    shared_ptr<BitArray> fac_restriction = nullptr;
+    if (py::extract<PyBA> (ael_restriction).check())
+      el_restriction = py::extract<PyBA>(ael_restriction)();
+
+    if (py::extract<PyBA> (afac_restriction).check())
+      fac_restriction = py::extract<PyBA>(afac_restriction)();
+
+    auto biform = make_shared<Rbfi_TT> (fes, aname, el_restriction, fac_restriction, flags);
+    biform -> SetCheckUnused (check_unused);
+    return biform;
+  }),
+  py::arg("space"),
+  py::arg("name") = "bfa",
+  py::arg("element_restriction") = DummyArgument(),
+  py::arg("facet_restriction") = DummyArgument(),
+  py::arg("check_unused") = true,
+  rblf_string_T)
+.def(py::init([](shared_ptr<FESpace> fes1,
+   shared_ptr<FESpace> fes2,
+   const string & aname,
+   py::object ael_restriction,
+   py::object afac_restriction,
+   bool check_unused,
[…]
