### 2021-05-27 82c349c
RestrictedBilinearFormDouble instead of monkey patching
