+++ python/ngs_check.py
@@ -1,4 +1,4 @@
-__ngsolve_required__ = "6.2.2102-119"
+__ngsolve_required__ = "6.2.2102-211"
 
 def check_if_ngsolve_newer_than(ngsolve_version_required):
     """
