### 2021-06-14 395268a
New Timer interface in NGSolve
Timer and RegionTimer are thread-safe
-> ThreadRegionTimer is removed
