### 2021-06-17 cfa98fd
explicit template function instantiation (needed after ngs update for clang)
