### 2021-07-02 2c0bfcb
update timer handling due to changes in NGSolve
