+++ lsetcurving/shiftintegrators.cpp
@@ -42,9 +42,9 @@ namespace ngfem
                                                 LocalHeap & lh,
                                                 shared_ptr<LsetEvaluator<D>> lseteval) const
   {
-    static Timer timer ("ShiftIntegrator<D>::CalcElementVector",1); 
+    static Timer timer ("ShiftIntegrator<D>::CalcElementVector"); 
     RegionTracer regt (TaskManager::GetThreadId(),timer);
-    ThreadRegionTimer reg (timer,TaskManager::GetThreadId());
+    NgProfiler::RegionTimer reg (timer);
     
     elvec = 0.0;
     const ScalarFiniteElement<D> & scafe = dynamic_cast<const ScalarFiniteElement<D> &>(fel);
