+++ lsetcurving/calcpointshift.cpp
@@ -151,7 +151,7 @@ namespace ngfem
     double * n_maxits)
   {
     static int timer = NgProfiler::CreateTimer ("SearchCorrespondingPoint"); 
-    ThreadRegionTimer reg (timer,TaskManager::GetThreadId());
+    NgProfiler::RegionTimer reg (timer);
     // static Timer time_not_conv ("SearchCorrespondingPoint::not converged");
     // static Timer time_conv ("SearchCorrespondingPoint::converged");
     // static Timer time_its ("SearchCorrespondingPoint::iterations");
