+++ cutint/spacetimecutrule.cpp
@@ -102,9 +102,9 @@ namespace xintegration
                                                         int order_space,
                                                         SWAP_DIMENSIONS_POLICY quad_dir_policy,
                                                         LocalHeap & lh){
-        static Timer timer("SpaceTimeCutIntegrationRule",1);
+        static Timer timer("SpaceTimeCutIntegrationRule");
         RegionTracer rt(TaskManager::GetThreadId(), timer);
-        ThreadRegionTimer reg(timer,TaskManager::GetThreadId());
+        NgProfiler::RegionTimer reg(timer);
         //cout << "This is SpaceTimeCutIntegrationRule " << endl;
         ELEMENT_TYPE et_space = trafo.GetElementType();
         
