+++ xfem/symboliccutbfi.cpp
@@ -78,7 +78,7 @@ namespace ngfem
   {
 
     static int timer = NgProfiler::CreateTimer ("SymbolicCutBFI::CalcElementMatrixAdd");
-    ThreadRegionTimer reg (timer, TaskManager::GetThreadId());
+    NgProfiler::RegionTimer reg (timer);
 
     if (element_vb != VOL)
       {
@@ -707,8 +707,8 @@ namespace ngfem
 
     if (etfacet != ET_SEGM && lsetintdom->GetDomainType() == IF) // Codim 2 special case (3D -> 1D)
     {
-      static Timer t("symbolicCutBFI - CoDim2-hack", 2);
-      ThreadRegionTimer reg (t, TaskManager::GetThreadId());
+      static Timer t("symbolicCutBFI - CoDim2-hack", NoTracing);
+      NgProfiler::RegionTimer reg (t);
       static bool first = true;
       if (first)
       {
