### 2021-07-02 8b8f156
[ci] minor: switch from schruste/ngsolve to ngsxfem/ngsolve
