+++ demos/unf_interf_prob.py
@@ -203,8 +203,16 @@ Draw(lsetp1, mesh, "levelset_P1")
 # Note that standard netgen-gui visualization does not respect
 # the discontinuities, they are smeared out. To see kinks or jumps
 # more clearly increase the subdivision option in the visualization.
-DrawDC(lsetp1, gfu.components[0], gfu.components[1],
-       mesh, "u", deformation=lsetadap.deform)
+if formulation == "CUTFEM":
+  DrawDC(lsetp1, gfu.components[0], gfu.components[1],
+         mesh, "u", deformation=lsetadap.deform)
+  Draw(IfPos(-lsetp1, alpha[0]*grad(gfu.components[0]), alpha[1]*grad(gfu.components[1])),
+         mesh, "sigma", deformation=lsetadap.deform)
+else:
+  DrawDC(lsetp1, gfu.components[0]+neg(gfu.components[1]), gfu.components[0]+pos(gfu.components[1]),
+         mesh, "u", deformation=lsetadap.deform)
+  Draw(IfPos(-lsetp1, alpha[0]*grad(gfu.components[0])+neg_grad(gfu.components[1]), alpha[1]*grad(gfu.components[0])+pos_grad(gfu.components[1])),
+         mesh, "sigma", deformation=lsetadap.deform)
 
 # Computation of L2 error:
 err_sqr = sum([(gfuh[i] - solution[i])**2 * dx[i].order(2 * order)
