### 2021-07-03 b8bdc53
[ci] minor: fix typo (nsg -> ngs)
