### 2021-07-03 e99cc95
[demos/unf_ifp] fix pycodestyle issues
