### 2021-07-07 #3237 @meg-simula
Hi @schruste,

Thanks again for the submission and for handling all reviewer
feedback. I assume that the paper in the master branch is the relevant
one.

Could you address the following editorial points:
- [ ] Use Python, not python.
- [ ] Spell out w.r.t. (with regard to)
- [ ] Replace "facilities", perhaps by "functionality"?
- [ ] Use .~ to avoid breaking white space (e.g.~e.g.)
- [ ] Replace "advanced features beyond that:", by e.g. "advanced features beyond these, including:".

After addressing these points, could you:
