### 2021-07-07 d1ddcfd
[paper.md] changes for joss
