### 2021-07-07 #3237 @schruste
@meg-simula . Thank you for your reply!

Indeed the latest changes have been made to the master branch.

I applied your changes, but have problems with "Use .~ to avoid breaking white space (e.g.~e.g.)" as this doesn't behave in markdown the same way one would expect in latex, the tilde is rendered.
