### 2021-07-09 #3237 @schruste
Ok, I think we are done:
- [x] we fixed the issue with the whitespaces.
- [x] The release tag is [`joss`](https://github.com/ngsxfem/ngsxfem/releases/tag/joss)
- [x] This version is now on zenodo with DOI 10.5281/zenodo.5081124, [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5081124.svg)](https://doi.org/10.5281/zenodo.5081124)
