+++ CMakeLists.txt
@@ -105,7 +105,8 @@ endif(SPACETIME_SANITY_CHECKS)
 find_package(NGSolve REQUIRED CONFIG)
 set(NGSXFEM_VERSION_MAJOR 2)
 set(NGSXFEM_VERSION_MINOR 0)
-set(NGSXFEM_VERSION "${NGSXFEM_VERSION_MAJOR}.${NGSXFEM_VERSION_MINOR}")
+set(NGSXFEM_VERSION_CYCLE 2104)
+set(NGSXFEM_VERSION "${NGSXFEM_VERSION_MAJOR}.${NGSXFEM_VERSION_MINOR}.${NGSXFEM_VERSION_CYCLE}")
 set(PACKAGE_VERSION ${NGSXFEM_VERSION})
 
 
