+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM ngsxfem/ngsolve:v6.2.2101-143
+FROM ngsxfem/ngsolve:v6.2.2104
 
 ARG NB_USER=jovyan
 ARG NB_UID=1000
@@ -6,12 +6,26 @@ ENV USER ${NB_USER}
 ENV NB_UID ${NB_UID}
 ENV HOME /home/${NB_USER}
 
+USER root
+RUN apt-get install -y cmake g++
 USER ${NB_USER}
         
 WORKDIR ${HOME}
 
-RUN git clone -b v2.0 --single-branch https://github.com/ngsxfem/ngsxfem.git ngsxfem
-RUN pip3 install git+https://github.com/ngsxfem/ngsxfem.git@v2.0 --user --upgrade --verbose
+## manual build based on local directory:
+#RUN mkdir ngsxfem
+#COPY . ${HOME}/ngsxfem/
+#USER root
+#RUN chown -R ${NB_UID} ${HOME}/ngsxfem
+#USER ${NB_USER}
+
+#WORKDIR ${HOME}/ngsxfem
+#RUN ls -al
+#RUN pip3 install . --user --upgrade --verbose
+
+## build based on github release:
+RUN git clone -b v2.0.2104 --single-branch https://github.com/ngsxfem/ngsxfem.git ngsxfem
+RUN pip3 install git+https://github.com/ngsxfem/ngsxfem.git@v2.0.2104 --user --upgrade --verbose
                 
 RUN python3 -c "import ngsolve; import xfem"        
                 
