### 2021-07-29 eed8d27
pytest to check that dCut/dFacetpatch return identical results to SymbilicBFI/FacePatchBFI. See #89
(cherry picked from commit ac611af802f473f18b9d5d9831738ffd2aaa2025)
