+++ tests/pytests/test_differential_symbol.py
@@ -1,0 +1,167 @@
+import pytest
+from netgen.geom2d import unit_square
+from ngsolve import *
+from xfem import *
+from xfem.lsetcurv import *
+ngsglobals.msg_level = 0
+
+mesh = Mesh(unit_square.GenerateMesh(maxh=0.2))
+levelset = (x - 0.5)**2 + (y - 0.5)**2 - 0.33**2
+
+
+@pytest.mark.parametrize('order', [1, 2, 3])
+@pytest.mark.parametrize('DOM', [POS, IF, NEG])
+@pytest.mark.parametrize('subdivlvl', [0, 1, 2])
+
+def test_cut_symbols_straightcut(order, DOM, subdivlvl):
+
+    if order > 1 and subdivlvl == 0:
+        lsetmeshadap = LevelSetMeshAdaptation(mesh, order=order,
+                                              levelset=levelset)
+        lset_h = lsetmeshadap.lset_p1
+    elif order == 1 and subdivlvl == 0:
+        lsetmeshadap = NoDeformation(mesh, levelset)
+        lset_h = lsetmeshadap.lset_p1
+    elif subdivlvl > 0:
+        lsetmeshadap = NoDeformation(mesh, levelset)
+        lset_h = GridFunction(H1(mesh, order))
+        lset_h.Set(levelset)
+
+    deform = lsetmeshadap.deform
+
+    ci = CutInfo(mesh, lset_h)
+    els_hasdom = ci.GetElementsOfType(HAS(DOM))
+    facests_none = BitArray(mesh.nedge)
+    facests_none.Clear()
+
+    V = H1(mesh, order=order)
+    u, v = V.TnT()
+
+    gfu = GridFunction(V)
+    gfu.Set(sin(x))
+
+    w1, w2 = gfu.vec.CreateVector(), gfu.vec.CreateVector()
+
+    # Differential Symbol Version
+    dx = dCut(lset_h, DOM, subdivlvl=subdivlvl, definedonelements=els_hasdom,
+              deformation=deform)
+
+    a1 = RestrictedBilinearForm(V, element_restriction=els_hasdom,
+                                facet_restriction=facests_none,
+                                check_unused=False)
+    a1 += Grad(u) * Grad(v) * dx
+    a1 += u * v * dx
+    a1.Assemble()
+    w1.data = a1.mat * gfu.vec
+
+    # SymbolicBFI version
+    lset_dom = {'levelset': lset_h, 'domain_type': DOM, 'subdivlvl': subdivlvl}
+    a2 = RestrictedBilinearForm(V, element_restriction=els_hasdom,
+                                facet_restriction=facests_none,
+                                check_unused=False)
+    a2 += SymbolicBFI(levelset_domain=lset_dom, form=Grad(u) * Grad(v),
+                      definedonelements=els_hasdom)
+    a2 += SymbolicBFI(levelset_domain=lset_dom, form=u * v,
+                      definedonelements=els_hasdom)
+    mesh.SetDeformation(deform)
+    a2.Assemble()
+    mesh.UnsetDeformation()
+    w2.data = a2.mat * gfu.vec
+
+    # Check
+    w1.data -= w2
+    diff = Norm(w1)
+    print(f'diff : {diff}')
+    assert diff < 1e-12
+
+
+@pytest.mark.parametrize('order', [1, 2, 3])
+@pytest.mark.parametrize('DOM', [POS, NEG])
+
+def test_cut_symbols_facetpatch(order, DOM):
+
+    # Higher order level set approximation
+    if order > 1:
+        lsetmeshadap = LevelSetMeshAdaptation(mesh, order=order,
+                                              levelset=levelset)
[…]
