### 2021-07-30 8097fe1
fix GetFacetsWithNeighborTypes for non-periodic facets
