### 2021-07-30 ec325e8
[CI] let build-related stuff run only on c5 (for now)
