+++ .gitlab-ci.yml
@@ -4,6 +4,8 @@ stages:
     - deploy
         
 build:
+  tags:
+    - c5
   stage: build
   image: ngsxfem/ngsolve:latest
   script:
@@ -19,6 +21,8 @@ build:
       - install
 
 ctests-demos:
+  tags:
+    - c5
   stage: test
   image: ngsxfem/ngsolve:latest
   script:
@@ -27,6 +31,8 @@ ctests-demos:
   allow_failure: false
 
 ctests-mayfail:
+  tags:
+    - c5
   stage: test
   image: ngsxfem/ngsolve:latest
   script:
@@ -35,6 +41,8 @@ ctests-mayfail:
   allow_failure: true
 
 ctests-pytests:
+  tags:
+    - c5
   stage: test
   image: ngsxfem/ngsolve:latest
   script:
