### 2021-07-30 2a1bf10
pytest cutsymbols -> add test for dg integrals, see #89
(cherry picked from commit e3a556f9071cadd949ccbd2ff837299226138f20)
