### 2021-08-09 #3237 @meg-simula
@schruste Thanks again for the submission and the revisions! The paper will now undergo a quick check by the JOSS Editors-in-Chief before final acceptance. No need to do anything at your end at this point.
